Bugfix. Setting coordinates if papersize changes.
Changing the papersize has an effect on the  coordinates of the printed map.
If the last action before creating the print is the change of papersize the coordinates are not set to the correct values.
This leads to a gap between walkingpapers an data layer.
So it is  needed to call the onMapChanged(map) callback after change of papersize.